Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240119/합과 평균/sum-and-avg.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main() {
     double a, b;
     cin >> a >> b;
-    cout << a+b << " " << (a+b)/2;
+    cout << a+b << " " << (double)(a+b)/2;
     // 여기에 코드를 작성해주세요.
     return 0;
 }
